+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/easyadapter/chat/ChatActivityView.kt
@@ -5,6 +5,7 @@ import android.text.format.DateFormat
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.activity_recycler.*
+import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.sticky.StickyHeaderDecor
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.ItemList
 import ru.surfstudio.android.recycler.decorator.Decorator
@@ -13,7 +14,6 @@ import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.controll
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.controller.MessageTimeController
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.ChatDecorOffset
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.ChatMessageDecor
-import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.MessageTimeDecor
 import ru.surfstudio.android.recycler.decorator.easyadapter.offset
 import ru.surfstudio.android.recycler.decorator.easyadapter.underlay
 
@@ -40,8 +40,8 @@ class ChatActivityView : AppCompatActivity() {
         }
 
         val decorator = Decorator.Builder()
-                .underlay(messageTimeController.viewType() to MessageTimeDecor(this))
                 .underlay(chatController.viewType() to ChatMessageDecor(this))
+                .overlay(StickyHeaderDecor())
                 .offset(chatController.viewType() to ChatDecorOffset())
                 .build()