+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/easyadapter/chat/decor/sticky/StiсkyHolder.kt
@@ -1,0 +1,7 @@
+package ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.sticky
+
+/**
+ * Mark interface for [StickyHeaderDecor]
+ */
+interface StiсkyHolder {
+}