+++ recycler-decorator/RELEASE_NOTES.md
@@ -14,3 +14,4 @@
 * fix gradle config for recycler-decorator-easyadapter
 * Fix readme
 * ANDDEP-1048 Fixing wrong docs links and docs structure
+* add sample for sticky header decor